Relaxed type interfaces
