+++ src/main/scala/com/johnsnowlabs/nlp/embeddings/WordEmbeddingsWriter.scala
@@ -1,6 +1,6 @@
 package com.johnsnowlabs.nlp.embeddings
 
-import java.nio.ByteBuffer
+import java.nio.{ByteBuffer, ByteOrder}
 
 import com.johnsnowlabs.storage.{RocksDBConnection, StorageWriter}
 
@@ -8,10 +8,14 @@ class WordEmbeddingsWriter(
                             connection: RocksDBConnection,
                             autoFlushAfter: Option[Integer] = None
                           )
-  extends StorageWriter[Float](connection, autoFlushAfter) {
+  extends StorageWriter[Array[Float]](connection, autoFlushAfter) {
 
-  override protected def addToBuffer(buffer: ByteBuffer, content: Float): Unit = {
-    buffer.putFloat(content)
+  override def toBytes(content: Array[Float]): Array[Byte] = {
+    val buffer = ByteBuffer.allocate(content.length * 4)
+    buffer.order(ByteOrder.LITTLE_ENDIAN)
+    for (value <- content) {
+      buffer.putFloat(value)
+    }
+    buffer.array()
   }
-
 }
